# backend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ApiPageCommentView" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/CommentView"
            },
            "type" : "array"
          },
          "nextCursor" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiPageUUID" : {
        "properties" : {
          "items" : {
            "items" : {
              "format" : "uuid",
              "type" : "string"
            },
            "type" : "array"
          },
          "nextCursor" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "CommentView" : {
        "properties" : {
          "authorId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "body" : {
            "type" : "string"
          },
          "commentId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "postId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "CreateCommentRequest" : {
        "properties" : {
          "body" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "FollowRelationship" : {
        "properties" : {
          "followedByViewer" : {
            "type" : "boolean"
          },
          "followerCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "followingCount" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "FollowRelationshipView" : {
        "properties" : {
          "followedByViewer" : {
            "type" : "boolean"
          },
          "followerCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "followingCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "userId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "PostCommentsView" : {
        "properties" : {
          "commentCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "postId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ProblemDetail" : {
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "properties" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "format" : "uri",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "The HTTP API the Pictogram SPA drives.",
    "title" : "Pictogram API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/comments" : {
      "get" : {
        "operationId" : "countsByPostIds",
        "parameters" : [ {
          "in" : "query",
          "name" : "postIds",
          "required" : true,
          "schema" : {
            "items" : {
              "format" : "uuid",
              "type" : "string"
            },
            "type" : "array",
            "uniqueItems" : true
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/PostCommentsView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "The comment count for each requested post."
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The request asked for more ids than the batch limit."
          }
        },
        "tags" : [ "comments-controller" ]
      }
    },
...
    "/api/follows" : {
      "get" : {
        "operationId" : "relationshipsByIds",
        "parameters" : [ {
          "in" : "query",
          "name" : "ids",
          "required" : true,
          "schema" : {
            "items" : {
              "format" : "uuid",
              "type" : "string"
            },
            "type" : "array",
            "uniqueItems" : true
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/FollowRelationshipView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "The viewer's follow standing with each requested user."
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The request asked for more ids than the batch limit."
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The caller has no valid access token."
          }
        },
        "tags" : [ "follow-controller" ]
      }
    },
    "/api/follows/{userId}" : {
      "delete" : {
        "operationId" : "unfollow",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "The viewer no longer follows the user (or never did)."
          }
        },
        "tags" : [ "follow-controller" ]
      },
      "get" : {
        "operationId" : "relationship",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FollowRelationship"
                }
              }
            },
            "description" : "The user's follower / following counts and the viewer's relationship."
          }
        },
        "tags" : [ "follow-controller" ]
      },
      "put" : {
        "operationId" : "follow",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "The viewer now follows the user (or already did)."
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The viewer tried to follow themselves."
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The viewer is following too fast."
          }
        },
        "tags" : [ "follow-controller" ]
      }
    },
    "/api/follows/{userId}/followers" : {
      "get" : {
        "operationId" : "followers",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiPageUUID"
                }
              }
            },
            "description" : "A page of the users who follow this user, newest follow first."
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The caller has no valid access token."
          }
        },
        "tags" : [ "follow-controller" ]
      }
    },
    "/api/follows/{userId}/following" : {
      "get" : {
        "operationId" : "following",
        "parameters" : [ {
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiPageUUID"
                }
              }
            },
            "description" : "A page of the users this user follows, newest follow first."
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The caller has no valid access token."
          }
        },
        "tags" : [ "follow-controller" ]
      }
    },
...
    "/api/posts/{postId}/comments" : {
      "get" : {
        "operationId" : "thread",
        "parameters" : [ {
          "in" : "path",
          "name" : "postId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiPageCommentView"
                }
              }
            },
            "description" : "A page of the post's comments, oldest first."
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The pagination cursor is malformed."
          }
        },
        "tags" : [ "comment-controller" ]
      },
      "post" : {
        "operationId" : "add",
        "parameters" : [ {
          "in" : "path",
          "name" : "postId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateCommentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommentView"
                }
              }
            },
            "description" : "The comment was added.",
            "headers" : {
              "Location" : {
                "description" : "The comment's URL.",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The comment is empty or longer than 1000 characters."
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The caller has no valid access token."
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The viewer is commenting too fast."
          }
        },
        "tags" : [ "comment-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "url" : "/"
  } ]
}